React Virtual. Int

+++ buildSrc/src/main/kotlin/karakum/virtual/Converter.kt
@@ -51,7 +51,7 @@ private fun convertConst(
     val body = source.replace(": {", " {")
         .replace("\n    ", "\n    val ")
         .replace("<any>", "<*>")
-        .replace("number", "Double")
+        .replace("number", "Int")
 
     val content = "external $type $body"
     return ConversionResult(name, content)
